refactor(relayer): refactor the utils files
